Rename: photo booth raw 이름 변경
- raw-data.repository.ts -> hidden.repository.ts 변경

Related to #4
