+++ src/photo-booth/photo-booth.module.ts
@@ -5,7 +5,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { PhotoBooth } from './entity/photo-booth.entity';
 import { HiddenPhotoBooth } from './entity/raw-data.entity';
 import { PhotoBoothRepository } from './repository/photo-booth.repository';
-import { HiddenBoothRepository } from './repository/photo-booth-raw-data.repository';
+import { HiddenBoothRepository } from './repository/photo-booth-hidden.repository';
 import {
   Hashtag,
   PhotoBoothBrand,
