+++ src/photo-booth/photo-booth.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { PhotoBoothService } from './photo-booth.service';
 import { PhotoBoothRepository } from './repository/photo-booth.repository';
-import { HiddenBoothRepository } from './repository/photo-booth-raw-data.repository';
+import { HiddenBoothRepository } from './repository/photo-booth-hidden.repository';
 import { PhotoBoothBrandRepository } from './repository/photo-booth-brand.repository';
 import { PhotoBooth } from './entity/photo-booth.entity';
 import { GetPhotoBoothListDto } from './dto/get-photo-booth-list.dto';
