+++ src/photo-booth/repository/photo-booth-hidden.repository.ts
